Split classes delivery_token and idelivery_token
